【操作】 功能-开发 【概述】 文章组列表查询功能开发 【编号】 无 【依赖模块】 无 【备注】无

+++ src/main/java/com/cccloud/blogs/easybatis/interfaces/Page.java
@@ -1,1 +1,1 @@
-package com.cccloud.blogs.easybatis.interfaces;/** * 创建人：徐卫超 * 时间：2019/6/20 9:51 * 功能：分页接口 * 备注： */public interface Page {    String START = "pageStart";    String OFFSET = "pageOffset";    Integer getPageStart();    Integer getPageOffset();}
+package com.cccloud.blogs.easybatis.interfaces;/** * 创建人：徐卫超 * 时间：2019/6/20 9:51 * 功能：分页接口 * 备注： */public interface Page {    String START = "pageStart";    String OFFSET = "pageOffset";    long getPageStart();    long getPageOffset();}
